backend/backend/requirements.txt
asgiref==3.6.0
autopep8==2.0.2
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.7
djangorestframework==3.14.0
idna==3.4
Pillow==9.4.0
psycopg2-binary==2.9.5
pycodestyle==2.10.0
python-dotenv==1.0.0
pytz==2023.2
requests==2.28.2
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.15
whitenoise==6.4.0
